ch[#161497422] set up database Add migration.js for creating and dropping tables Add productController.js for productController Restructure folder structure
