+++ UI/assets/css/lib/table.css
@@ -1,154 +1,25 @@
 @import url('https://fonts.googleapis.com/css?family=Josefin+Sans');
 
-.table {
-    display: table;
-    width: 100%;
-    background: #fdfdfd;
-    margin: 0;
-    box-sizing: border-box;
-    font-family: 'Josefin Sans', sans-serif;
-    font-size: 1.1em;
+table {
+  width: 100%;
+  max-width: 100%;
+  margin-top: 150px;
+  overflow-x:scroll;
 }
 
-.caption {
-    display: block;
-    width: 99.1%;
-    background: #4C4E5E;
-    height: 35px;
-    padding-left: 10px;
-    color: #fff;
-    font-size: 20px;
-    line-height: 35px;
-   
+th {
+  background: #262734;
+  color: #fff;
+  padding: 0.3em;
+  font-size: 1.4em;
+  font-weight: 100;
 }
 
-.header-row {
-    background: #262734;
-    color: #fff;
+tr:nth-child(even) {
+  background: #a7a7a7;
 }
 
-.row {
-    display: table-row;
-}
-
-.row:nth-child(even) {
-    background: #f1f1f1;
-}
-
-.cell {
-    display: table-cell;
-    padding: 6px;
-    text-align: center;
-    height: 35px;
-    padding-top: 24px;
-}
-
-.primary {
-    text-align: left;
-}
-
-
-input[type="radio"],
-input[type="checkbox"] {
-    display: none;
-}
-
-
-@media only screen and (max-width: 1000px) {
-
-    body {
-        padding: 0;
-    }
-
-    #table {
-        display: block;
-        margin: 44px 0 0 0;
-    }
-
-    .caption {
-        position: fixed;
-        top: 0;
-        text-align: center;
-        height: 44px;
-        line-height: 44px;
-        z-index: 5;
-        border-bottom: 2px solid #999;
-    }
-
-    .row {
-        position: relative;
-        border-bottom: 1px solid #ccc;
-
-    }
-
-    .cell {
-        display: block;
-        border: none;
-        position: relative;
-        height: 65px;
-        line-height: 45px;
-        text-align: left;
-    }   
-
-
-    input[type="radio"],
-    input[type="checkbox"] {
-        display: block;
-        position: absolute;
-        z-index: 1;
-        width: 99%;
-        height: 100%;
-        opacity: 0;
-    }
-
-    input[type="radio"]:checked,
-    input[type="checkbox"]:checked {
-        z-index: -1;
-    }
-
-    input[type="radio"]:checked~.cell,
-    input[type="checkbox"]:checked~.cell {
-        display: block;
-
-        border-bottom: 1px solid #eee;
-    }
-
-    input[type="radio"]:checked~.cell:nth-of-type(n+2),
-    input[type="checkbox"]:checked~.cell:nth-of-type(n+2) {
-
-        background: #e0e0e0;
-    }
-
-    input[type="radio"]:checked~.cell:nth-of-type(n+2):before,
-    input[type="checkbox"]:checked~.cell:nth-of-type(n+2):before {
-        content: attr(data-label);
-
-        display: inline-block;
-        width: 60px;
-        background: #999;
-        border-radius: 10px;
-        height: 20px;
-        margin-right: 10px;
-        font-size: 12px;
-        line-height: 20px;
-        text-align: center;
-        color: white;
-
-    }
-
-    input[type="radio"]:checked~.primary,
-    input[type="checkbox"]:checked~.primary {
-        border-bottom: 2px solid #999;
-    }
-
-    input[type="radio"]:checked~.primary:after,
-    input[type="checkbox"]:checked~.primary:after {
-        position: absolute;
-        right: 18px;
-        top: 22px;
-        border-right: 10px solid transparent;
-        border-left: 10px solid transparent;
-        border-top: 10px solid #ccc;
-        z-index: 2;
-    }
+td {
+  padding: 20px;
+  font-size: 1.3em;
 }
