+++ package.json
@@ -39,12 +39,15 @@
     "http-errors": "~1.6.2",
     "istanbul": "^0.4.5",
     "jsonwebtoken": "^8.3.0",
+    "make-runnable": "^1.3.6",
     "mocha": "^5.2.0",
+    "moment": "^2.22.2",
     "morgan": "~1.9.0",
     "nodemon": "^1.17.5",
     "nyc": "^13.1.0",
     "path": "^0.12.7",
-    "pg": "^7.5.0"
+    "pg": "^7.5.0",
+    "uuid": "^3.3.2"
   },
   "standard": {
     "env": [
