+++ server/models/db.js
@@ -10,4 +10,22 @@ const pool = new Pool({
   ssl,
 });
 
-export default pool;
+/**
+ * DB Query
+ * @param {object} req
+ * @param {object} res
+ * @returns {object} object
+ */
+const query = (text, params) => {
+  return new Promise((resolve, reject) => {
+    pool.query(text, params)
+      .then((res) => {
+        resolve(res);
+      })
+      .catch((err) => {
+        reject(err);
+      });
+  });
+};
+
+export { pool, query };
